I don't know, fixing merge problems?
I tried my best to fix push origin merge problems. I deleted unnecessary tile assets that weren't assigned sprites. Made the unity project RubysAdventureGuillenP2 in my GitHub repository to merge with the clone of it in my local files. The major problems that occurred were with conflicting files that the same name but different base/ID data. This happened with MainScene.unity and EditorUserSettings.asset and defaul-2021.dwlt that is the best of what I could describe what I did in this update/changes log.